Refactor COG test code
converters -> cog for geotiff_metadata
