+++ docs/api.rst
@@ -281,5 +281,4 @@ odc.geo.converters
    extract_gcps_raw
    map_crs
    rio_geobox
-   geotiff_metadata
